finall permissions with api

+++ app/Http/Controllers/API/GroupController.php
@@ -17,9 +17,26 @@ class GroupController extends Controller
      */
     public function index()
     {
-        $groups = Group::paginate();
+        $groups = Group::with('users')->whereId(auth()->id())->paginate();
         return response()->json($groups);
     }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
+    {
+        if (Group::whereHas('users', function ($q){
+                $q->where('user_id', auth()->id());
+            })->whereHas('permissions', function ($q){
+                $q->where('name', 'group_make');
+            })->count() == 0)
+            return response()->json(['permission' => 'false']);
+        return response()->json(['permission' => 'true']);
+    }
+
 
     /**
      * Store a newly created resource in storage.
@@ -48,6 +65,23 @@ class GroupController extends Controller
     }
 
     /**
+     * Show the form for editing the specified resource.
+     *
+     * @param int $id
+     * @return \Illuminate\Http\Response
+     */
+    public function edit($id)
+    {
+        if (Group::whereHas('users', function ($q){
+                $q->where('user_id', auth()->id());
+            })->whereHas('permissions', function ($q){
+                $q->where('name', 'group_edit');
+            })->count() == 0)
+            return response()->json(['permission' => 'false']);
+        return response()->json(['permission' => 'true']);
+    }
+
+    /**
      * Update the specified resource in storage.
      *
      * @param \Illuminate\Http\Request $request
@@ -70,9 +104,16 @@ class GroupController extends Controller
      */
     public function destroy($id)
     {
+        if (Group::whereHas('users', function ($q){
+                $q->where('user_id', auth()->id());
+            })->whereHas('permissions', function ($q){
+                $q->where('name', 'group_delete');
+            })->count() == 0)
+            return response()->json(['permission' => 'false']);
         $group = Group::find($id);
         $group->users()->detach();
+        $group->permissions()->detach();
         $group->delete();
-        return response()->json('Group was deleted');
+        return response()->json(['message' => 'Group was deleted']);
     }
 }
